split keycloakAuth into separate file
