+++ api/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -12,7 +11,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Nerzal/gocloak/v8"
 	_ "github.com/joho/godotenv/autoload"
 )
 
@@ -191,7 +189,7 @@ func getContactSegmentsById(w http.ResponseWriter, r *http.Request, contactId st
 		} else if err != nil {
 			fmt.Fprintf(w, "Decode failed with error %s\n", err)
 		}
-		// Get contact ID
+		// Mark contact segments as true
 		for key := range data.Lists {
 			contactSegments[key] = true
 		}
@@ -238,58 +236,4 @@ func updateContactSegments(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-}
-
-// keycloak authentication function that wraps handlers needing keycloak auth
-func keycloakAuth(fn func(http.ResponseWriter, *http.Request)) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Add("Content-Type", "application/json")
-		w.Header().Add("Access-Control-Allow-Origin", "*")
-		w.Header().Add("Access-Control-Allow-Credentials", "true")
-		w.Header().Add("Access-Control-Allow-Headers", "Content-Type, Authorization, Email, ContactId, SegmentsAndIds")
-		w.Header().Add("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE")
-
-		if r.Method == "OPTIONS" {
-			http.Error(w, "No Content", http.StatusNoContent)
-			return
-		}
-		authHeader := strings.Fields(r.Header.Get("Authorization"))
-
-		if len(authHeader) < 2 {
-			w.WriteHeader(http.StatusBadRequest)
-			fmt.Fprintf(w, "Invalid authorization header")
-			return
-		}
-		token := authHeader[1]
-		kcClientID := os.Getenv("KC_CLIENT_ID")
-		kcClientSecret := os.Getenv("KC_CLIENT_SECRET")
-		kcRealm := os.Getenv("KC_REALM")
-		kcURL := os.Getenv("KC_URL")
-		kcClient := gocloak.NewClient((kcURL))
-
-		ctx := context.Background()
-		_, err := kcClient.LoginClient(ctx, kcClientID, kcClientSecret, kcRealm)
-
-		if err != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			fmt.Fprintf(w, "Keycloak Login failed:"+err.Error())
-			return
-		}
-
-		rptResult, err := kcClient.RetrospectToken(ctx, token, kcClientID, kcClientSecret, kcRealm)
-		if err != nil {
-			w.WriteHeader(http.StatusUnauthorized)
-			fmt.Fprintf(w, "Keycloak inspection failed:"+err.Error())
-			return
-		}
-
-		if !*rptResult.Active {
-			w.WriteHeader(http.StatusUnauthorized)
-			fmt.Fprintf(w, "Keycloak token is not active")
-			return
-		}
-
-		//Execute handler function if token is valid
-		fn(w, r)
-	}
 }
